requirements.txt
streamlit==1.32.0
google-generativeai==0.5.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1